requirements.txt
fastapi
uvicorn[standard]
starlette
Jinja2
python-jose
passlib
pymongo[srv]